+++ core/module/user/view/profilEdit/profilEdit.php
@@ -101,201 +101,24 @@
         </div>
     </div>
     <div class="containerModule">
-        <div class="row">
-            <div class="col12">
-                <div class="block">
-                    <h4>
-                        <?php echo helper::translate('Permissions sur le module') . ' ' . helper::translate('Blog'); ?>
-                    </h4>
-                    <div class="row">
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditBlogAdd', true, 'Ajouter', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'add'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditBlogEdit', true, 'Éditer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'edit'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditBlogDelete', true, 'Effacer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'delete'])
-                            ]); ?>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditBlogOption', true, 'Options', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'option'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditBlogComment', true, 'Gérer les commentaires', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'comment'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditBlogCommentApprove', true, 'Approuver les commentaires', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'commentApprove'])
-                            ]); ?>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="col6">
-                            <?php echo template::checkbox('profilEditBlogCommentDelete', true, 'Effacer les commentaires', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'commentDelete'])
-                            ]); ?>
-                        </div>
-                        <div class="col6">
-                            <?php echo template::checkbox('profilEditBlogCommentDeleteAll', true, 'Nettoyer les commentaires', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'commentDeleteAll'])
-                            ]); ?>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col12">
-                <div class="block">
-                    <h4>
-                        <?php echo helper::translate('Permissions sur le module') . ' ' . helper::translate('News'); ?>
-                    </h4>
-                    <div class="row">
-                        <div class="col3">
-                            <?php echo template::checkbox('profilEditNewsAdd', true, 'Ajouter', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'news', 'add'])
-                            ]); ?>
-                        </div>
-                        <div class="col3">
-                            <?php echo template::checkbox('profilEditNewsEdit', true, 'Éditer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'news', 'edit'])
-                            ]); ?>
-                        </div>
-                        <div class="col3">
-                            <?php echo template::checkbox('profilEditNewsDelete', true, 'Effacer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'news', 'delete'])
-                            ]); ?>
-                        </div>
-                        <div class="col3">
-                            <?php echo template::checkbox('profilEditNewsOption', true, 'Options', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'news', 'option'])
-                            ]); ?>
-                        </div>
-
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col12">
-                <div class="block">
-                    <h4>
-                        <?php echo helper::translate('Permissions sur le module') . ' ' . helper::translate('Galerie'); ?>
-                    </h4>
-                    <div class="row">
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditGalleryAdd', true, 'Ajouter', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'gallery', 'add'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditGalleryEdit', true, 'Éditer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'gallery', 'edit'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditGalleryDelete', true, 'Effacer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'gallery', 'delete'])
-                            ]); ?>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="col6">
-                            <?php echo template::checkbox('profilEditGalleryOption', true, 'Options', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'gallery', 'option'])
-                            ]); ?>
-                        </div>
-                        <div class="col6">
-                            <?php echo template::checkbox('profilEditGalleryTheme', true, 'Thème', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'gallery', 'theme'])
-                            ]); ?>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col12">
-                <div class="block">
-                    <h4>
-                        <?php echo helper::translate('Permissions sur le module') . ' ' . helper::translate('Formulaire'); ?>
-                    </h4>
-                    <div class="row">
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditFormOption', true, 'Options', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'form', 'option'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditFormData', true, 'Gérer les données', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'form', 'data'])
-                            ]); ?>
-                        </div>
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditFormExport2csv', true, 'Export CSV', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'form', 'export2csv'])
-                            ]); ?>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="col6">
-                            <?php echo template::checkbox('profilEditFormDelete', true, 'Effacer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'form', 'delete'])
-                            ]); ?>
-                        </div>
-                        <div class="col6">
-                            <?php echo template::checkbox('profilEditFormDeleteAll', true, 'Tout Effacer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'form', 'deleteAll'])
-                            ]); ?>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col12">
-                <div class="block">
-                    <h4>
-                        <?php echo helper::translate('Permissions sur le module') . ' ' . helper::translate('Redirection'); ?>
-                    </h4>
-                    <div class="row">
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditRedirectionConfig', true, 'Configurer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'redirection', 'config'])
-                            ]); ?>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col12">
-                <div class="block">
-                    <h4>
-                        <?php echo helper::translate('Permissions sur le module') . ' ' . helper::translate('Recherche'); ?>
-                    </h4>
-                    <div class="row">
-                        <div class="col4">
-                            <?php echo template::checkbox('profilEditSearchConfig', true, 'Configurer', [
-                                'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'search', 'config'])
-                            ]); ?>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
+        <?php if (file_exists('module/blog/profil/view/edit.inc.php')) {
+            include('module/blog/profil/view/edit.inc.php');
+        } ?>
+        <?php if (file_exists('module/news/profil/view/edit.inc.php')) {
+            include('module/news/profil/view/edit.inc.php');
+        } ?>
+        <?php if (file_exists('module/gallery/profil/view/edit.inc.php')) {
+            include('module/gallery/profil/view/edit.inc.php');
+        } ?>
+        <?php if (file_exists('module/form/profil/view/edit.inc.php')) {
+            include('module/form/profil/view/edit.inc.php');
+        } ?>
+        <?php if (file_exists('module/redirection/profil/view/edit.inc.php')) {
+            include('module/redirection/profil/view/edit.inc.php');
+        } ?>
+        <?php if (file_exists('module/search/profil/view/edit.inc.php')) {
+            include('module/search/profil/view/edit.inc.php');
+        } ?>
     </div>
 <?php endif; ?>
 <div class="row">