+++ module/blog/profil/view/edit.inc.php
@@ -1,0 +1,55 @@
+<div class="row">
+    <div class="col12">
+        <div class="block">
+            <h4>
+                <?php echo helper::translate('Blog'); ?>
+            </h4>
+            <div class="row">
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditBlogAdd', true, 'Ajouter', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'add'])
+                    ]); ?>
+                </div>
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditBlogEdit', true, 'Éditer', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'edit'])
+                    ]); ?>
+                </div>
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditBlogDelete', true, 'Supprimer', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'delete'])
+                    ]); ?>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditBlogOption', true, 'Options', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'option'])
+                    ]); ?>
+                </div>
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditBlogComment', true, 'Gérer les commentaires', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'comment'])
+                    ]); ?>
+                </div>
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditBlogCommentApprove', true, 'Approuver les commentaires', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'commentApprove'])
+                    ]); ?>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col6">
+                    <?php echo template::checkbox('profilEditBlogCommentDelete', true, 'Supprimer les commentaires', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'commentDelete'])
+                    ]); ?>
+                </div>
+                <div class="col6">
+                    <?php echo template::checkbox('profilEditBlogCommentDeleteAll', true, 'Nettoyer les commentaires', [
+                        'checked' => $this->getData(['profil', $this->getUrl(2), $this->getUrl(3), 'blog', 'commentDeleteAll'])
+                    ]); ?>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>