+++ module/redirection/profil/view/add.inc.php
@@ -1,0 +1,14 @@
+<div class="row">
+    <div class="col12">
+        <div class="block">
+            <h4>
+                <?php echo helper::translate('Redirection'); ?>
+            </h4>
+            <div class="row">
+                <div class="col4">
+                    <?php echo template::checkbox('profilEditRedirectionConfig', true, 'Configurer'); ?>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>