+++ module/form/changes.md
@@ -1,3 +1,5 @@
+# Version 4.0
+- Gestion des permissions intégrée dans le module
 # Version 3.10
 - Bloque l'effacement selon le profil
 - Masque le code de vérification