+++ module/redirection/changes.md
@@ -1,2 +1,4 @@
+# Version 2.0
+- Gestion des permissions intégrée dans le module
 # Version 1.6
 - Multilinguisme